SCARV: Fix data type widths for xc.pclmul.*
 On branch dev/xcrypto
 Changes to be committed:
	modified:   ../riscv/insns/xc_pclmul_h.h
	modified:   ../riscv/insns/xc_pclmul_l.h

 Changes not staged for commit:
	modified:   ../riscv/decode.h
	modified:   ../riscv/riscv.mk.in

 Untracked files:
	../riscv/insns/xc_macc_1.h
	../riscv/insns/xc_macc_2.h
	../riscv/insns/xc_madd_2.h
	../riscv/insns/xc_madd_3.h
	../riscv/insns/xc_mclmul_3.h
	../riscv/insns/xc_mequ.h
	../riscv/insns/xc_mgte.h
	../riscv/insns/xc_mlte.h
	../riscv/insns/xc_mmul_3.h
	../riscv/insns/xc_msll.h
	../riscv/insns/xc_msll_i.h
	../riscv/insns/xc_msrl.h
	../riscv/insns/xc_msrl_i.h
	../riscv/insns/xc_msub_2.h
	../riscv/insns/xc_msub_3.h

+++ riscv/insns/xc_pclmul_h.h
@@ -13,9 +13,9 @@ uint32_t result     = 0;
 
 for(uint32_t i = 0; i < n; i ++) {
 
-    uint32_t lhs    = (XCRS1 >> (i*w)) & mask;
-    uint32_t rhs    = (XCRS2 >> (i*w)) & mask;
-    uint32_t pr     = (XC_CARRYLESS_MUL(lhs,rhs) >> w) & mask;
+    uint64_t lhs    = (XCRS1 >> (i*w)) & mask;
+    uint64_t rhs    = (XCRS2 >> (i*w)) & mask;
+    uint64_t pr     = (XC_CARRYLESS_MUL(lhs,rhs) >> w) & mask;
     
     result         |= pr << (i*w);
 }
